applying resolvers structure
