+++ Retos/sesion3-reto1/RoggerValverde/db/autores.json
@@ -1,1 +1,1 @@
-[{"id":-411296849,"nombres":"Rogger","apellidos":"Valverde","nacionalidad":"venezolano"},{"id":838980758,"nombres":"Oscar","apellidos":"Ramirez"},{"id":-690879222,"nombres":"Jorge","apellidos":"Flores"},{"id":626178308,"nombres":"Oscar","apellidos":"Ramirez2"},{"id":1381337490,"nombres":"Oscar","apellidos":"Ramirez3"},{"id":-869196751,"nombres":"Oscar","apellidos":"Ramirez4","nacionalidad":"gg"}]
+[{"id":-411296849,"nombres":"Rogger","apellidos":"Valverde","nacionalidad":"venezolano"},{"id":838980758,"nombres":"Oscar","apellidos":"Ramirez"},{"id":-690879222,"nombres":"Jorge","apellidos":"Flores"},{"id":626178308,"nombres":"Oscar","apellidos":"Ramirez2"},{"id":1381337490,"nombres":"Oscar","apellidos":"Ramirez3"},{"id":-869196751,"nombres":"Oscar","apellidos":"Ramirez4","nacionalidad":"gg"},{"id":-1006863094,"nombres":"Rogger Andre","apellidos":"Valverde","nacionalidad":"peruana"}]
